requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
faiss-cpu
sentence-transformers
redis
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
tiktoken
pypdf2